success build android apk
1. modify some server code to complie under linux
2. modify some build cfg and code to make success build android apk
3. can use DDMS see log

+++ server/ServerBase/MySqlManager.h
@@ -20,7 +20,7 @@ public:
 	template<typename Func>
 	bool Execute(const char *sql, Func callbackFunc, void *param, s32 len)
 	{
-		if( len > DB_PARAM_MAX_LEN )
+		if( len > (s32)DB_PARAM_MAX_LEN )
 		{
 			return false;
 		}
@@ -41,7 +41,7 @@ public:
 	template<class ClassObj, class ClassMemFunc>
 	bool Execute(const char *sql, const ClassObj &classObj, ClassMemFunc memCallbackFunc, void *param, s32 len)
 	{
-		if( len > DB_PARAM_MAX_LEN )
+		if( len > (s32)DB_PARAM_MAX_LEN )
 		{
 			return false;
 		}
